* Load AI Agents, agent is a uint16 and should be loaded as one, thanks voodoo * some work on shared_ptr debugging, a lot less spam now, disabled Aura debugging until I fix it * clear tmpAura on Aura::Remove()
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/Hearthstone@1001 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/ObjectStorage.cpp
@@ -159,7 +159,7 @@ void ObjectMgr::LoadExtraCreatureProtoStuff()
 			{
 				Field *fields = result->Fetch();
 				entry = fields[0].GetUInt32();
-				agent = fields[1].GetUInt32();
+				agent = fields[1].GetUInt16();
 				spellID = fields[4].GetUInt32();
 				int32 tcd = fields[7].GetInt32();
